添加websoket连接

+++ datetype/client.go
@@ -1,9 +1,9 @@
 package datetype
 
 type MemoryData struct {
-	Total       uint64  `json:"total"`
-	Used        uint64  `json:"used"`
-	Free        uint64  `json:"free"`
+	Total       float64 `json:"total"`
+	Used        float64 `json:"used"`
+	Free        float64 `json:"free"`
 	UsedPercent float64 `json:"usedPercent"`
 }
 type CPUData struct {
@@ -14,18 +14,18 @@ type CPUData struct {
 }
 type DiskData struct {
 	Mountpoint  string  `json:"mountpoint"`
-	Total       uint64  `json:"total"`        // 总大小，单位字节
-	Used        uint64  `json:"used"`         // 已使用，单位字节
-	Free        uint64  `json:"free"`         // 可用，单位字节
+	Total       float64 `json:"total"`        // 总大小，单位字节
+	Used        float64 `json:"used"`         // 已使用，单位字节
+	Free        float64 `json:"free"`         // 可用，单位字节
 	UsedPercent float64 `json:"used_percent"` // 使用率，百分比
 }
 
 type NetworkData struct {
-	Name        string `json:"name"`
-	BytesSent   uint64 `json:"bytesSent"`
-	BytesRecv   uint64 `json:"bytesRecv"`
-	PacketsSent uint64 `json:"packetsSent"`
-	PacketsRecv uint64 `json:"packetsRecv"`
+	Name        string  `json:"name"`
+	BytesSent   float64 `json:"bytesSent"`
+	BytesRecv   float64 `json:"bytesRecv"`
+	PacketsSent float64 `json:"packetsSent"`
+	PacketsRecv float64 `json:"packetsRecv"`
 }
 
 type MonitorData struct {
